# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Torrent streaming server.",
        "title": "Swagger Torrserver API",
        "contact": {},
        "license": {
            "name": "GPL 3.0"
        },
        "version": "{version.Version}"
    },
    "basePath": "/",
    "paths": {
...
        "/play/{hash}/{id}": {
            "get": {
                "description": "Play given torrent referenced by hash.",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "API"
                ],
                "summary": "Play given torrent referenced by hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Torrent hash",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "File index in torrent (-10 for auto-selection)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Season number for auto-selection (requires id=-10)",
                        "name": "season",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Episode number for auto-selection (requires id=-10)",
                        "name": "episode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filename for auto-selection (requires id=-10)",
                        "name": "filename",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Torrent data"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}